use 32 bit integers

+++ src/treelearner/histogram_compressor.hpp
@@ -15,16 +15,16 @@ class HistogramCompressor {
   HistogramCompressor(const int num_threads);
 
   template <typename S_HIST_T, typename U_HIST_T>
-  void Compress(const S_HIST_T* in_buffer, uint8_t* out_buffer, uint8_t* out_bits_buffer, data_size_t num_bin, uint64_t* thread_total_half_bytes_offset);
+  void Compress(const S_HIST_T* in_buffer, uint8_t* out_buffer, uint8_t* out_bits_buffer, data_size_t num_bin, uint32_t* thread_total_half_bytes_offset);
 
   template <typename S_HIST_T, typename U_HIST_T>
-  void Decompress(const uint8_t* in_buffer, const uint8_t* in_bits_buffer, const uint64_t* thread_half_byte_offset, data_size_t num_bin, S_HIST_T* out_buffer);
+  void Decompress(const uint8_t* in_buffer, const uint8_t* in_bits_buffer, const uint32_t* thread_half_byte_offset, data_size_t num_bin, S_HIST_T* out_buffer);
 
   void Test();
 
  private:
   template <typename S_HIST_T, typename U_HIST_T>
-  uint64_t ComputeThreadHalfBytes(const S_HIST_T* in_buffer, uint8_t* out_bits_buffer,
+  uint32_t ComputeThreadHalfBytes(const S_HIST_T* in_buffer, uint8_t* out_bits_buffer,
     uint8_t* thread_first_bits_buffer,
     data_size_t start_bin, data_size_t end_bin);
 
@@ -35,7 +35,7 @@ class HistogramCompressor {
     uint8_t* thread_first_buffer,
     data_size_t start_bin,
     data_size_t end_bin,
-    uint64_t thread_start_half_bytes);
+    uint32_t thread_start_half_bytes);
 
   std::vector<uint8_t> thread_first_bits_;
   std::vector<uint8_t> thread_first_;
